Size the feature set for the dialect's whole inventory
`markdown_core_feature_set` was one bit per registry row in a 32-bit word,
while `docs/specs/dialect.md` lists 36 features beyond CommonMark for the
plan's later items to register; row 32 would have shifted past the word.
The set is now 64 bits wide, and the table's definition carries a C99
compile-time check that the row count fits, so outgrowing the set is a
build error rather than an unselectable feature.

+++ packages/markdown-core/extensions/feature-registry.h
@@ -46,8 +46,10 @@ typedef struct markdown_core_feature {
     int option_bit;
 } markdown_core_feature;
 
-/** A set of registry rows: bit `i` is the row `markdown_core_feature_at(i)`. */
-typedef uint32_t markdown_core_feature_set;
+/** A set of registry rows: bit `i` is the row `markdown_core_feature_at(i)`.
+ * Wide enough for every row the dialect specification lists, and the table
+ * is checked against this width where it is defined. */
+typedef uint64_t markdown_core_feature_set;
 
 size_t markdown_core_feature_count(void);
